+++ src/imio/helpers/tests/test_transmogrifier.py
@@ -4,7 +4,7 @@ from imio.helpers.testing import IntegrationTestCase
 from imio.helpers.transmogrifier import filter_keys
 from imio.helpers.transmogrifier import get_main_path
 from imio.helpers.transmogrifier import relative_path
-from imio.helpers.transmogrifier import text_int_to_bool
+from imio.helpers.transmogrifier import text_to_bool
 
 import os
 
@@ -38,16 +38,22 @@ class TestTesting(IntegrationTestCase):
         self.assertEquals(relative_path(self.portal, '/plone/directory'), 'directory')
         self.assertEquals(relative_path(self.portal, '/alone/directory'), '/alone/directory')
 
-    def test_text_int_to_bool(self):
-        dic = {'0': '0', 0: 0, '1': '1', 123: 123, 'a': 'a', 'errors': 0}
+    def test_text_to_bool(self):
+        dic = {'True': 'True', 'true': 'true', 'False': 'False', 'false': 'false',
+               '0': '0', 0: 0, '1': '1', 123: 123,
+               'a': 'a', 'errors': 0}
 
         def logger(item, msg):
             item['errors'] += 1
 
-        self.assertFalse(text_int_to_bool(dic, '0', logger))  # using fake log method to test error
-        self.assertFalse(text_int_to_bool(dic, 0, logger))
-        self.assertTrue(text_int_to_bool(dic, '1', logger))
-        self.assertTrue(text_int_to_bool(dic, 123, logger))
+        self.assertTrue(text_to_bool(dic, 'True', logger))
+        self.assertTrue(text_to_bool(dic, 'true', logger))
+        self.assertFalse(text_to_bool(dic, 'False', logger))
+        self.assertFalse(text_to_bool(dic, 'false', logger))
+        self.assertFalse(text_to_bool(dic, '0', logger))  # using fake log method to test error
+        self.assertFalse(text_to_bool(dic, 0, logger))
+        self.assertTrue(text_to_bool(dic, '1', logger))
+        self.assertTrue(text_to_bool(dic, 123, logger))
         self.assertEquals(dic['errors'], 0)
-        self.assertFalse(text_int_to_bool(dic, 'a', logger))
+        self.assertFalse(text_to_bool(dic, 'a', logger))
         self.assertEquals(dic['errors'], 1)